chore: rename ERC20 to CEP18
