+++ client-js/src/CEP18Client.ts
@@ -26,7 +26,7 @@ import {
 
 const { Contract } = Contracts;
 
-export default class ERC20Client {
+export default class CEP18Client {
   public contractClient: Contracts.Contract;
 
   constructor(public nodeAddress: string, public networkName: string) {
@@ -46,7 +46,7 @@ export default class ERC20Client {
   }
 
   /**
-   * Intalls ERC20
+   * Intalls CEP18
    * @param wasm contract representation of Uint8Array
    * @param args contract install arguments @see {@link InstallArgs}
    * @param paymentAmount payment amount required for installing the contract
@@ -249,21 +249,21 @@ export default class ERC20Client {
   }
 
   /**
-   * Returns the name of the ERC20 token.
+   * Returns the name of the CEP18 token.
    */
   public async name(): Promise<string> {
     return this.contractClient.queryContractData(['name']) as Promise<string>;
   }
 
   /**
-   * Returns the symbol of the ERC20 token.
+   * Returns the symbol of the CEP18 token.
    */
   public async symbol(): Promise<string> {
     return this.contractClient.queryContractData(['symbol']) as Promise<string>;
   }
 
   /**
-   * Returns the decimals of the ERC20 token.
+   * Returns the decimals of the CEP18 token.
    */
   public async decimals(): Promise<BigNumber> {
     return this.contractClient.queryContractData([
@@ -272,7 +272,7 @@ export default class ERC20Client {
   }
 
   /**
-   * Returns the total supply of the ERC20 token.
+   * Returns the total supply of the CEP18 token.
    */
   public async totalSupply(): Promise<BigNumber> {
     return this.contractClient.queryContractData([
