+++ client-js/src/index.ts
@@ -1,4 +1,4 @@
-export { default as ERC20Client } from './ERC20Client';
+export { default as CEP18Client } from './CEP18Client';
 export * from './error';
 export * from './types';
 export { default as ContractWASM } from './wasm';
